+++ src/ui/react/settings/components/header/index.tsx
@@ -63,7 +63,7 @@ export const Header = () => {
   const ExtraInfo = RouteExtraInfo[location.pathname];
 
   return (
-    <div className={cs.header} data-tauri-drag-region>
+    <div className={cs.header}>
       <div className={cs.title}>
         {label}
         {ExtraInfo && <ExtraInfo />}